test: tighten AC-5 diagnostic and dedupe minimal PDK fixture
Address Codex Round 2 review findings:

- Tighten test_check_pdk_overrides_nonexistent_tech to require the exact
  "PDK tech LEF not found" diagnostic instead of a generic validation
  heading disjunction, matching AC-5's contract
- Remove unused monkeypatch argument from test_check_pdk_overrides_valid
- Promote create_minimal_ics55_pdk and minimal_ics55_pdk_factory to
  test/conftest.py so test/data and test/cli share one canonical PDK
  layout definition via pytest fixture inheritance, removing the
  duplicated builder from both narrower conftests

Verification:
- 5 PDK-override CLI checks pass
- test/data/test_pdk.py: 21 passed
- Full non-integration suite: 1109 passed
- ruff check and format: clean
