+++ test/conftest.py
@@ -39,3 +39,27 @@ def pytest_collection_modifyitems(config, items):
             pdk_root = f"{repo_root}/{pdk_root}"
         if not complete_ics55_pdk_available(pdk_root):
             item.add_marker(skip_missing_pdk)
+
+
+def create_minimal_ics55_pdk(root):
+    root = Path(root)
+    tech_path = root / "prtech" / "techLEF" / "N551P6M_ecos.lef"
+    tech_path.parent.mkdir(parents=True, exist_ok=True)
+    tech_path.write_text("VERSION 5.8 ;\n")
+
+    stdcell_root = root / "IP" / "STD_cell" / "ics55_LLSC_H7C_V1p10C100"
+    for flavor in ("ics55_LLSC_H7CR", "ics55_LLSC_H7CL"):
+        lef_path = stdcell_root / flavor / "lef" / f"{flavor}_ecos.lef"
+        lef_path.parent.mkdir(parents=True, exist_ok=True)
+        lef_path.write_text("VERSION 5.8 ;\n")
+
+        lib_path = stdcell_root / flavor / "liberty" / f"{flavor}_ss_rcworst_1p08_125_nldm.lib"
+        lib_path.parent.mkdir(parents=True, exist_ok=True)
+        lib_path.write_text("library(test) { }\n")
+
+    return root
+
+
+@pytest.fixture
+def minimal_ics55_pdk_factory():
+    return create_minimal_ics55_pdk
